change the type of the specialTariff from float to double

+++ src/main/java/net/axel/models/entities/Contract.java
@@ -9,7 +9,7 @@ public class Contract {
     private UUID id;
     private Date startDate;
     private Date endDate;
-    private float specialTariff;
+    private double specialTariff;
     private String conditionsAccord;
     private Boolean renewable;
     private ContractStatus contractStatus;
